// apps/silo/src/apps/slack/types/workobjects.ts
export interface IIssueDetailService {
  getWorkItemDetails(
    planeClient: Client,
    ctx: TWorkItemDetailsCtx & { workspaceSlug: string }
  ): Promise<TWorkObjectIssueDetails>;
}

export interface IIssueWorkObjectViewService {
  getWorkItemView(
    workspaceSlug: string,
    issueDetails: TWorkObjectIssueDetails,
    config?: TWorkObjectViewConfig
  ): TWorkObjectView;
}

/* -------------------- TYPES -------------------------- */
export type TWorkObjectType = "MINIMAL" | "DETAILED"; // WHY NOT USE AN ENUM OVER HERE?

export enum EWorkObjectEntityType {
  FILE = "slack#/entities/file",
  TASK = "slack#/entities/task",
  ITEM = "slack#/entities/item",
  CONTENT_ITEM = "slack#/entities/content_item",
  INCIDENT = "slack#/entities/incident",
}

export type TWorkItemDetailsCtx =
  | { strategy: "sequence"; projectIdentifier: string; issueSequence: number }
  | { strategy: "id"; projectId: string; issueId: string };
// ...
export type TWorkObjectAdditionalIssueDetails = {
  propertyDetails?: {
    properties: ExIssueProperty[];
    propertyOptions: Map<string, ExIssuePropertyOption[]>;
    propertyValues: Map<string, { property_id: string; values: string[] }[]>;
  };
  availableStates?: ExState[];
  projectMembers?: PlaneUser[];

  // Issue Properties, that we need to show
  parent?: ExIssue;
};

export type TWorkObjectIssueDetails = IssueWithExpanded<
  ["state", "project", "assignees", "labels", "type", "created_by", "updated_by"]
> & {
  additionalDetails?: TWorkObjectAdditionalIssueDetails;
};